Refactor bigquery_util
into 3 smaller modules to make it's maintanance more manageable.
