+++ gcp_variant_transforms/libs/bigquery_util_test.py
@@ -481,85 +481,3 @@ class BigqueryUtilTest(unittest.TestCase):
                      bigquery_util.get_table_base_name(with_two_suffixes1))
     self.assertEqual(without_suffix2,
                      bigquery_util.get_table_base_name(with_two_suffixes2))
-
-  def test_calculate_optimal_range_interval(self):
-    range_end_to_expected_range_interval = {
-        39980000: 10000,
-        (39980000 - 1): 10000,
-        (39980000 - 9999): 10000,
-        39990000: 10010,
-        40000000: 10010,
-        40010000: 10010,
-        40020000: 10020,
-        40030000: 10020,
-    }
-    for range_end, expected_range_interval in (
-        range_end_to_expected_range_interval.items()):
-      (range_interval, range_end_enlarged) = (
-          bigquery_util.calculate_optimal_range_interval(range_end))
-      expected_range_end = (
-          expected_range_interval * (bigquery_util._MAX_BQ_NUM_PARTITIONS - 1))
-      self.assertEqual(expected_range_interval, range_interval)
-      self.assertEqual(expected_range_end, range_end_enlarged)
-
-  def test_calculate_optimal_range_interval_large(self):
-    large_range_ends = [bigquery_util.MAX_RANGE_END,
-                        bigquery_util.MAX_RANGE_END + 1,
-                        bigquery_util.MAX_RANGE_END - 1,
-                        bigquery_util.MAX_RANGE_END - 2 * pow(10, 4)]
-
-    expected_interval = int(bigquery_util.MAX_RANGE_END /
-                            float(bigquery_util._MAX_BQ_NUM_PARTITIONS))
-    expected_end = bigquery_util.MAX_RANGE_END
-    for large_range_end in large_range_ends:
-      (range_interval, range_end_enlarged) = (
-          bigquery_util.calculate_optimal_range_interval(large_range_end))
-      self.assertEqual(expected_interval, range_interval)
-      self.assertEqual(expected_end, range_end_enlarged)
-
-
-class FlattenCallColumnTest(unittest.TestCase):
-  """Test cases for class `FlattenCallColumn`."""
-
-  def setUp(self):
-    # We never query this table for running the following test, however, the
-    # mock values are based on this table's schema. In other words:
-    #   mock_columns.return_value = self._flatter._get_column_names()
-    #   mock_sub_fields.return_value = self._flatter._get_call_sub_fields()
-    input_base_table = ('gcp-variant-transforms-test:'
-                        'bq_to_vcf_integration_tests.'
-                        'merge_option_move_to_calls')
-    self._flatter = bigquery_util.FlattenCallColumn(input_base_table, ['chr20'])
-
-  @mock.patch('gcp_variant_transforms.libs.bigquery_util_test.bigquery_util.'
-              'FlattenCallColumn._get_column_names')
-  @mock.patch('gcp_variant_transforms.libs.bigquery_util_test.bigquery_util.'
-              'FlattenCallColumn._get_call_sub_fields')
-  def test_get_flatten_column_names(self, mock_sub_fields, mock_columns):
-    mock_columns.return_value = (
-        ['reference_name', 'start_position', 'end_position', 'reference_bases',
-         'alternate_bases', 'names', 'quality', 'filter', 'call', 'NS', 'DP',
-         'AA', 'DB', 'H2'])
-    mock_sub_fields.return_value = (
-        ['sample_id', 'genotype', 'phaseset', 'DP', 'GQ', 'HQ'])
-    expected_select = (
-        'main_table.reference_name AS `reference_name`, '
-        'main_table.start_position AS `start_position`, '
-        'main_table.end_position AS `end_position`, '
-        'main_table.reference_bases AS `reference_bases`, '
-        'main_table.alternate_bases AS `alternate_bases`, '
-        'main_table.names AS `names`, '
-        'main_table.quality AS `quality`, '
-        'main_table.filter AS `filter`, '
-        'call_table.sample_id AS `call_sample_id`, '
-        'call_table.genotype AS `call_genotype`, '
-        'call_table.phaseset AS `call_phaseset`, '
-        'call_table.DP AS `call_DP`, '
-        'call_table.GQ AS `call_GQ`, '
-        'call_table.HQ AS `call_HQ`, '
-        'main_table.NS AS `NS`, '
-        'main_table.DP AS `DP`, '
-        'main_table.AA AS `AA`, '
-        'main_table.DB AS `DB`, '
-        'main_table.H2 AS `H2`')
-    self.assertEqual(expected_select, self._flatter._get_flatten_column_names())
