+++ gcp_variant_transforms/transforms/sample_info_to_bigquery.py
@@ -72,7 +72,8 @@ class SampleInfoToBigQuery(beam.PTransform):
         both sample_name and file_name will be used.
     """
     self._output_table = bigquery_util.compose_table_name(
-        output_table_prefix, bigquery_util.SAMPLE_INFO_TABLE_SUFFIX)
+        output_table_prefix,
+        sample_info_table_schema_generator.SAMPLE_INFO_TABLE_SUFFIX)
     self._append = append
     self._sample_name_encoding = sample_name_encoding
     self._schema = sample_info_table_schema_generator.generate_schema()
